Move `transformLoopBody` to the new utils file
